Use new API in tests
